+++ packages/transaction-pay-controller/CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `TransactionPayController:setIsMaxAmount` messenger action to set max amount flag for a transaction ([#7562](https://github.com/MetaMask/core/pull/7562))
+- **BREAKING:** Support max amount quotes ([#7562](https://github.com/MetaMask/core/pull/7562))
+  - Add `TransactionPayController:setIsMaxAmount` messenger action.
   - Add `isMaxAmount` property to `TransactionData` type.
+  - Add `targetAmount` property to `TransactionPayQuote` and `TransactionPayTotals`.
   - Update Relay quote requests to use `EXACT_INPUT` trade type when max amount is selected.
   - Update totals calculation to account for max amount selection.